backend/worker/pyproject.toml
[project]
name = "docsearch-worker"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "celery[redis]>=5.4.0",
    "sqlalchemy[asyncio]>=2.0.0",
    "asyncpg>=0.29.0",
    "pymupdf>=1.24.0",
    "python-docx>=1.1.0",
    "openpyxl>=3.1.0",
    "pytesseract>=0.3.10",
    "sentence-transformers>=3.0.0",
    "boto3>=1.34.0",
    "qdrant-client>=1.9.0",
    "pydantic-settings>=2.6.0",
    "pillow>=10.0.0",
    "psycopg2-binary>=2.9.0",
]
